refactor(mcp): extract tool registrations into shared buildServer factory
Moves the McpServer construction and every server.tool(...) registration
out of mcp-server/index.js into mcp-server/lib-tools.mjs, exporting
buildServer({ url, apiKey, client, fetchImpl }). index.js keeps the CLI
dispatch untouched and now just calls buildServer() with env-sourced
config before connecting StdioServerTransport. Lets a future HTTP
transport reuse the same tools. Also adds lib-tools.mjs to package.json's
files array so the published package actually ships the module index.js
imports.
